When viewing a group without any users, ensure users list is empty.
This is a bit of a hack, that assumes there is no user with the ID of `0`. There shouldn't be, but who knows what some plugins might be doing. Maybe `PHP_INT_MAX` is a better idea, but we'll reserve that for later.

Fixes a bug that would default to showing all users if there were no users in a group.